requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.4.0
python-dotenv==1.0.1
email-validator==2.2.0
